requirements.txt
pillow>=10.0
opencv-python-headless>=4.8
scikit-image>=0.21
requests>=2.31
pyyaml>=6.0
ollama>=0.4
ultralytics>=8.0
